getSimilar implementations for in-memory "db"

+++ src/ow/directory/comparator/HammingIDComparator.java
@@ -8,18 +8,18 @@ import ow.id.ID;
 public class HammingIDComparator implements KeySimilarityComparator<ID> {
 
   @Override
-  public double similarity(ID key1, ID key2) {
+  public float similarity(ID key1, ID key2) {
     if (key1.getSize() != key2.getSize()) {
       throw new IllegalArgumentException("Keys must have equal sizes for similarity comparison");
     }
 
-    final double keySizeInBits = key1.getSize() * 8;
+    final float keySizeInBits = key1.getSize() * 8;
     if (keySizeInBits == 0) {
       // yay, paranoia!
       throw new IllegalArgumentException("Key size of zero is invalid.");
     }
 
-    final double hammingDistance =
+    final float hammingDistance =
         key1.toBigInteger().xor(key2.toBigInteger()).bitCount();
     return hammingDistance / keySizeInBits;
   }
